+++ installer/dbscript/dbscript-2.sql
@@ -4258,7 +4258,10 @@ VALUES ('apiv2_time_time_sheet_config', 'API-v2 Time - Time Sheet Start Day Conf
        ('apiv2_time_timesheet_comment', 'API-v2 Time - Timesheet comment', 1, 1, 1, 0),
        ('apiv2_time_config_time_format', 'API-v2 Time - Time Config Format', 1, 0, 0, 0),
        ('apiv2_time_validation_project_activity_name', 'API-v2 Time - Project Activity Name Validation', 1, 0, 0, 0),
-       ('apiv2_time_project_activity_unique_validation', 'API-v2 Time - Project Activity Unique Validation', 1, 0, 0, 0);
+       ('apiv2_time_project_activity_unique_validation', 'API-v2 Time - Project Activity Unique Validation', 1, 0, 0, 0),
+       ('apiv2_time_employee_timesheets', 'API-v2 Time - Employee Timesheets', 1, 0, 0, 0);
+
+
 
 SET @time_module_id := (SELECT `id` FROM ohrm_module WHERE name = 'time' LIMIT 1);
 
@@ -4275,6 +4278,7 @@ SET @apiv2_time_project_activity_unique_validation_data_group_id := (SELECT `id`
 SET @apiv2_time_timesheet_comment_data_group_id := (SELECT `id` FROM ohrm_data_group WHERE name = 'apiv2_time_timesheet_comment' LIMIT 1);
 SET @apiv2_time_config_time_format_data_group_id := (SELECT `id` FROM ohrm_data_group WHERE name = 'apiv2_time_config_time_format' LIMIT 1);
 SET @apiv2_time_validation_project_activity_name_data_group_id := (SELECT `id` FROM ohrm_data_group WHERE name = 'apiv2_time_validation_project_activity_name' LIMIT 1);
+SET @apiv2_time_employee_timesheets_data_group_id := (SELECT `id` FROM ohrm_data_group WHERE name = 'apiv2_time_employee_timesheets' LIMIT 1);
 
 INSERT INTO ohrm_api_permission (`api_name`, `module_id`, `data_group_id`)
 VALUES ('OrangeHRM\\Time\\Api\\TimeConfigPeriodAPI', @time_module_id, @apiv2_time_time_sheet_config_data_group_id),
@@ -4289,7 +4293,8 @@ VALUES ('OrangeHRM\\Time\\Api\\TimeConfigPeriodAPI', @time_module_id, @apiv2_tim
        ('OrangeHRM\\Time\\Api\\TimesheetCommentAPI', @time_module_id, @apiv2_time_timesheet_comment_data_group_id),
        ('OrangeHRM\\Time\\Api\\TimesheetTimeFormatAPI', @time_module_id, @apiv2_time_config_time_format_data_group_id),
        ('OrangeHRM\\Time\\Api\\ValidationProjectActivityNameAPI', @time_module_id, @apiv2_time_validation_project_activity_name_data_group_id),
-       ('OrangeHRM\\Time\\Api\\TimesheetProjectActivityUniqueValidationAPI', @time_module_id, @apiv2_time_project_activity_unique_validation_data_group_id);
+       ('OrangeHRM\\Time\\Api\\TimesheetProjectActivityUniqueValidationAPI', @time_module_id, @apiv2_time_project_activity_unique_validation_data_group_id),
+       ('OrangeHRM\\Time\\Api\\EmployeeTimesheetAPI', @time_module_id, @apiv2_time_employee_timesheets_data_group_id);
 
 INSERT INTO ohrm_user_role_data_group (`can_read`, `can_create`, `can_update`, `can_delete`, `self`, `data_group_id`, `user_role_id`)
 VALUES (1, 0, 1, 0, 0, @apiv2_time_time_sheet_config_data_group_id, @admin_role_id),
@@ -4315,7 +4320,9 @@ VALUES (1, 0, 1, 0, 0, @apiv2_time_time_sheet_config_data_group_id, @admin_role_
        (1, 0, 0, 0, 0, @apiv2_time_config_time_format_data_group_id, @admin_role_id),
        (1, 0, 0, 0, 0, @apiv2_time_config_time_format_data_group_id, @ess_role_id),
        (1, 0, 0, 0, 0, @apiv2_time_validation_project_activity_name_data_group_id, @admin_role_id),
-       (1, 0, 0, 0, 0, @apiv2_time_validation_project_activity_name_data_group_id, @project_admin_role_id);
+       (1, 0, 0, 0, 0, @apiv2_time_validation_project_activity_name_data_group_id, @project_admin_role_id),
+       (1, 0, 0, 0, 0, @apiv2_time_employee_timesheets_data_group_id, @admin_role_id),
+       (1, 0, 0, 0, 0, @apiv2_time_employee_timesheets_data_group_id, @supervisor_role_id);
 
 UPDATE `ohrm_screen` SET `module_id`= 5  WHERE `action_url` = 'viewCustomers';
 UPDATE `ohrm_screen` SET `module_id`= 5  WHERE `action_url` = 'viewProjects';
