+++ symfony/plugins/orangehrmTimePlugin/test/fixtures/EmployeeTimesheetAPITest.yml
@@ -1,0 +1,49 @@
+Config:
+  - { name: 'authorize_user_role_manager_class', value: 'BasicUserRoleManager' }
+  - { name: 'timesheet_period_set', value: 'Yes' }
+  - { name: 'timesheet_period_and_start_date', value: '<TimesheetPeriod><PeriodType>Weekly</PeriodType><ClassName>WeeklyTimesheetPeriod</ClassName><StartDate>1</StartDate><Heading>Week</Heading></TimesheetPeriod>' }
+
+Employee:
+  - { empNumber: 1, firstName: 'Kayla', lastName: 'Abbey', middleName: 'T', employeeId: 'E001', emp_work_email: 'kayla@xample.com', emp_oth_email: 'kayla2@xample.com' }
+  - { empNumber: 2, firstName: 'Ashley', lastName: 'Abel', middleName: 'ST',employeeId: 'E002', emp_work_email: 'ashley@xample.com', emp_oth_email: 'ashley2@xample.com' }
+  - { empNumber: 3, firstName: 'Renukshan', lastName: 'Saputhanthri', middleName: 'RT', employeeId: 'E003', emp_work_email: 'renukshan@xample.com', emp_oth_email: 'renukshan2@xample.com' }
+  - { empNumber: 4, firstName: 'Sherin', lastName: 'Sherin', middleName: 'Sherin', employeeId: 'E004', emp_work_email: 'sherin@xample.com', emp_oth_email: 'sherin@xample.com' }
+
+ReportingMethod:
+  - { "reporting_method_id": "1","reporting_method_name": "Direct" }
+  - { "reporting_method_id": "2","reporting_method_name": "Indirect" }
+
+ReportTo:
+  - { "erep_sup_emp_number": "3","erep_sub_emp_number": "2","erep_reporting_mode": "1" }
+
+UserRole:
+  - { id: 1, name: Admin, display_name: Admin, is_assignable: 1, is_predefined: 1 }
+  - { id: 2, name: ESS, display_name: ESS, is_assignable: 1, is_predefined: 1 }
+  - { id: 3, name: Supervisor, display_name: Supervisor, is_assignable: 0, is_predefined: 1 }
+
+User:
+  - { "id": 1, "user_role_id": 1, "emp_number": 1, "user_name": "admin","user_password": "admin","deleted": "0","status": "1","date_modified": "2021-09-15 16:51:16","modified_user_id": "1" }
+  - { "id": 2, "user_role_id": 2, "emp_number": 2, "user_name": "Adalwin","user_password": "Adalwin","deleted": "0","status": "1","date_entered": "2021-09-15 16:47:14","created_by": "1" }
+  - { "id": 3, "user_role_id": 3, "emp_number": 3, "user_name": "Anderson","user_password": "Anderson","deleted": "0","status": "1","date_entered": "2021-09-15 16:47:53","created_by": "1" }
+  - { "id": 4, "user_role_id": 2, "emp_number": 4, "user_name": "Sherin","user_password": "Sherin","deleted": "0","status": "1","date_entered": "2021-09-15 17:47:59","created_by": "1" }
+
+Customer:
+  - { name: 'Orange', description: 'HRM', deleted: false }
+  - { name: 'CUS', description: 'DES', deleted: false }
+  - { name: 'CUS2', description: 'DES2', deleted: false }
+
+Project:
+  - { project_id: 1, customer_id: 1, name: 'OrangeHRM', description: 'firstproject', is_deleted: 0 }
+  - { project_id: 2, customer_id: 2, name: 'OrangeHRM2', description: 'secondproject', is_deleted: 0 }
+  - { project_id: 3, customer_id: 3, name: 'OrangeHRM3', description: 'thirdproject', is_deleted: 1 }
+
+ProjectActivity:
+  - { activity_id: 1, project_id: 1, name: 'Activity1 For Pro1', is_deleted: 0 }
+  - { activity_id: 2, project_id: 1, name: 'Activity2 For Pro1', is_deleted: 1 }
+
+Timesheet:
+  - { timesheet_id: 1, state: 'CREATED', start_date: '2011-04-18',end_date: '2011-04-24', employee_id: 1 }
+  - { timesheet_id: 2, state: 'SUBMITTED', start_date: '2011-05-22',end_date: '2011-05-29', employee_id: 2 }
+  - { timesheet_id: 3, state: 'ACCEPTED', start_date: '2011-04-18',end_date: '2011-04-24', employee_id: 3 }
+  - { timesheet_id: 4, state: 'NOT SUBMITTED', start_date: '2011-06-10',end_date: '2011-06-16', employee_id: 2 }
+  - { timesheet_id: 5, state: 'APPROVED', start_date: '2011-07-10',end_date: '2011-07-16', employee_id: 4 }
