+++ symfony/plugins/orangehrmTimePlugin/test/fixtures/testcases/EmployeeTimesheetAPITestCases.yaml
@@ -1,0 +1,226 @@
+GetAll:
+  'Get employee timesheets - self':
+    userId: 2
+    services:
+      time.timesheet_service: \OrangeHRM\Time\Service\TimesheetService
+    attributes:
+      empNumber: 2
+    query:
+      fromDate: '2011-05-22'
+      toDate: '2011-05-29'
+    body: [ ]
+    data:
+      - id: 2
+        status:
+          id: "SUBMITTED"
+          name: "Submitted"
+        startDate: "2011-05-22"
+        endDate: "2011-05-29"
+    meta:
+      total: 1
+
+  'Get employee timesheets - limit':
+    userId: 1
+    services:
+      time.timesheet_service: \OrangeHRM\Time\Service\TimesheetService
+    attributes:
+      empNumber: 1
+    query:
+      fromDate: '2011-04-18'
+      toDate: '2011-05-29'
+      limit: 1
+    body: [ ]
+    data:
+      - id: 1
+        status:
+          id: "CREATED"
+          name: "Created"
+        startDate: "2011-04-18"
+        endDate: "2011-04-24"
+    meta:
+      total: 1
+
+    Get employee timesheets - date param only':
+      userId: 1
+      services:
+        time.timesheet_service: \OrangeHRM\Time\Service\TimesheetService
+        core.config_service: \OrangeHRM\Core\Service\ConfigService
+      attributes:
+        empNumber: 1
+      query:
+        date: '2011-04-19'
+      body: [ ]
+      data:
+        - id: 1
+          status:
+            id: "CREATED"
+            name: "Created"
+          startDate: "2011-04-18"
+          endDate: "2011-04-24"
+      meta:
+        total: 1
+
+    'Get employee timesheets - no time sheet':
+      userId: 1
+      services:
+        time.timesheet_service: \OrangeHRM\Time\Service\TimesheetService
+      attributes:
+        empNumber: 2
+      query:
+        fromDate: "2021-11-29"
+        toDate: "2021-12-05"
+      body: [ ]
+      data: [ ]
+      meta:
+        total: 0
+
+    'Get employee timesheets - fromDate only':
+      userId: 1
+      services:
+        time.timesheet_service: \OrangeHRM\Time\Service\TimesheetService
+      attributes:
+        empNumber: 1
+      query:
+        fromDate: "2021-11-20"
+      body: [ ]
+      exception:
+        class: '\OrangeHRM\Core\Api\V2\Exception\BadRequestException'
+        message: 'To Date is required'
+
+    'Get employee timesheets - toDate only':
+      userId: 1
+      services:
+        time.timesheet_service: \OrangeHRM\Time\Service\TimesheetService
+      attributes:
+        empNumber: 1
+      query:
+        toDate: "2021-11-20"
+      exception:
+        class: '\OrangeHRM\Core\Api\V2\Exception\BadRequestException'
+        message: 'From Date is required'
+
+    'Get employee timesheets - fromDate, toDate and Date':
+      userId: 1
+      services:
+        time.timesheet_service: \OrangeHRM\Time\Service\TimesheetService
+      attributes:
+        empNumber: 1
+      query:
+        fromDate: "2021-11-29"
+        toDate: "2021-12-05"
+        date: "2021-11-30"
+      body: [ ]
+      exception:
+        class: '\OrangeHRM\Core\Api\V2\Exception\BadRequestException'
+        message: "You can't pass date param together with fromDate and toDate params"
+
+    'Get employee timesheets - fromDate >= toDate':
+      userId: 1
+      services:
+        time.timesheet_service: \OrangeHRM\Time\Service\TimesheetService
+      attributes:
+        empNumber: 1
+      query:
+        fromDate: "2021-11-20"
+        toDate: "2021-11-14"
+      body: [ ]
+      exception:
+        class: '\OrangeHRM\Core\Api\V2\Exception\BadRequestException'
+        message: "From date should be earlier than to date"
+
+    'Get employee timesheets - without params':
+      userId: 1
+      now:
+        datetime: '2011-04-19'
+      services:
+        time.timesheet_service: \OrangeHRM\Time\Service\TimesheetService
+        core.config_service: \OrangeHRM\Core\Service\ConfigService
+      attributes:
+        empNumber: 1
+      query:
+      body: [ ]
+      data:
+        - id: 1
+          status:
+            id: "CREATED"
+            name: "Created"
+          startDate: "2011-04-18"
+          endDate: "2011-04-24"
+      meta:
+        total: 1
+
+  'Get employee timesheets - Admin to Employee':
+    userId: 1
+    services:
+      time.timesheet_service: \OrangeHRM\Time\Service\TimesheetService
+    attributes:
+      empNumber: 2
+    query:
+      fromDate: '2011-05-22'
+      toDate: '2011-05-29'
+    body: [ ]
+    data:
+      - id: 2
+        status:
+          id: "SUBMITTED"
+          name: "Submitted"
+        startDate: "2011-05-22"
+        endDate: "2011-05-29"
+    meta:
+      total: 1
+
+  'Get employee timesheets - wrong emp number':
+    userId: 1
+    services:
+      time.timesheet_service: \OrangeHRM\Time\Service\TimesheetService
+    attributes:
+      empNumber: 20
+    query:
+      fromDate: '2011-05-22'
+      toDate: '2011-05-29'
+    body: [ ]
+    invalidOnly: [ 'empNumber' ]
+
+  'Get employee timesheets - Supervisor to subordinate':
+    userId: 3
+    services:
+      time.timesheet_service: \OrangeHRM\Time\Service\TimesheetService
+    attributes:
+      empNumber: 2
+    query:
+      fromDate: '2011-05-22'
+      toDate: '2011-05-29'
+    body: [ ]
+    data:
+      - id: 2
+        status:
+          id: "SUBMITTED"
+          name: "Submitted"
+        startDate: "2011-05-22"
+        endDate: "2011-05-29"
+    meta:
+      total: 1
+
+  'Get employee timesheets - Supervisor to non subordinate':
+    userId: 3
+    services:
+      time.timesheet_service: \OrangeHRM\Time\Service\TimesheetService
+    attributes:
+      empNumber: 4
+    query:
+      fromDate: '2011-05-22'
+      toDate: '2011-05-29'
+    body: [ ]
+    invalidOnly: [ 'empNumber' ]
+
+  'Get employee timesheets - ESS to ESS':
+    userId: 2
+    services:
+      time.timesheet_service: \OrangeHRM\Time\Service\TimesheetService
+    attributes:
+      empNumber: 4
+    query:
+      fromDate: '2011-05-22'
+      toDate: '2011-05-29'
+    body: [ ]
+    invalidOnly: [ 'empNumber' ]
